test: extract clicking tests
