+++ test/input.spec.js
@@ -16,121 +16,18 @@
 
 const path = require('path');
 const utils = require('./utils');
-const DeviceDescriptors = utils.requireRoot('DeviceDescriptors');
-const iPhone = DeviceDescriptors['iPhone 6'];
 
 module.exports.addTests = function({testRunner, expect}) {
   const {describe, xdescribe, fdescribe} = testRunner;
   const {it, fit, xit} = testRunner;
   const {beforeAll, beforeEach, afterAll, afterEach} = testRunner;
   describe('input', function() {
-    it('should click the button', async({page, server}) => {
-      await page.goto(server.PREFIX + '/input/button.html');
-      await page.click('button');
-      expect(await page.evaluate(() => result)).toBe('Clicked');
-    });
-    it('should click with disabled javascript', async({page, server}) => {
-      await page.setJavaScriptEnabled(false);
-      await page.goto(server.PREFIX + '/wrappedlink.html');
-      await Promise.all([
-        page.click('a'),
-        page.waitForNavigation()
-      ]);
-      expect(page.url()).toBe(server.PREFIX + '/wrappedlink.html#clicked');
-    });
-
-    it('should click offscreen buttons', async({page, server}) => {
-      await page.goto(server.PREFIX + '/offscreenbuttons.html');
-      const messages = [];
-      page.on('console', msg => messages.push(msg.text()));
-      for (let i = 0; i < 11; ++i) {
-        // We might've scrolled to click a button - reset to (0, 0).
-        await page.evaluate(() => window.scrollTo(0, 0));
-        await page.click(`#btn${i}`);
-      }
-      expect(messages).toEqual([
-        'button #0 clicked',
-        'button #1 clicked',
-        'button #2 clicked',
-        'button #3 clicked',
-        'button #4 clicked',
-        'button #5 clicked',
-        'button #6 clicked',
-        'button #7 clicked',
-        'button #8 clicked',
-        'button #9 clicked',
-        'button #10 clicked'
-      ]);
-    });
-
-    it('should click wrapped links', async({page, server}) => {
-      await page.goto(server.PREFIX + '/wrappedlink.html');
-      await Promise.all([
-        page.click('a'),
-        page.waitForNavigation()
-      ]);
-      expect(page.url()).toBe(server.PREFIX + '/wrappedlink.html#clicked');
-    });
-
-    it('should click on checkbox input and toggle', async({page, server}) => {
-      await page.goto(server.PREFIX + '/input/checkbox.html');
-      expect(await page.evaluate(() => result.check)).toBe(null);
-      await page.click('input#agree');
-      expect(await page.evaluate(() => result.check)).toBe(true);
-      expect(await page.evaluate(() => result.events)).toEqual([
-        'mouseover',
-        'mouseenter',
-        'mousemove',
-        'mousedown',
-        'mouseup',
-        'click',
-        'input',
-        'change',
-      ]);
-      await page.click('input#agree');
-      expect(await page.evaluate(() => result.check)).toBe(false);
-    });
-
-    it('should click on checkbox label and toggle', async({page, server}) => {
-      await page.goto(server.PREFIX + '/input/checkbox.html');
-      expect(await page.evaluate(() => result.check)).toBe(null);
-      await page.click('label[for="agree"]');
-      expect(await page.evaluate(() => result.check)).toBe(true);
-      expect(await page.evaluate(() => result.events)).toEqual([
-        'click',
-        'input',
-        'change',
-      ]);
-      await page.click('label[for="agree"]');
-      expect(await page.evaluate(() => result.check)).toBe(false);
-    });
-
-    it('should fail to click a missing button', async({page, server}) => {
-      await page.goto(server.PREFIX + '/input/button.html');
-      let error = null;
-      await page.click('button.does-not-exist').catch(e => error = e);
-      expect(error.message).toBe('No node found for selector: button.does-not-exist');
-    });
-    // @see https://github.com/GoogleChrome/puppeteer/issues/161
-    it('should not hang with touch-enabled viewports', async({page, server}) => {
-      await page.setViewport(iPhone.viewport);
-      await page.mouse.down();
-      await page.mouse.move(100, 10);
-      await page.mouse.up();
-    });
     it('should type into the textarea', async({page, server}) => {
       await page.goto(server.PREFIX + '/input/textarea.html');
 
       const textarea = await page.$('textarea');
       await textarea.type('Type in this text!');
       expect(await page.evaluate(() => result)).toBe('Type in this text!');
-    });
-    it('should click the button after navigation ', async({page, server}) => {
-      await page.goto(server.PREFIX + '/input/button.html');
-      await page.click('button');
-      await page.goto(server.PREFIX + '/input/button.html');
-      await page.click('button');
-      expect(await page.evaluate(() => result)).toBe('Clicked');
     });
     it('should upload the file', async({page, server}) => {
       await page.goto(server.PREFIX + '/input/fileupload.html');
@@ -280,43 +177,6 @@ module.exports.addTests = function({testRunner, expect}) {
       expect(newDimensions.width).toBe(width + 104);
       expect(newDimensions.height).toBe(height + 104);
     });
-    it('should scroll and click the button', async({page, server}) => {
-      await page.goto(server.PREFIX + '/input/scrollable.html');
-      await page.click('#button-5');
-      expect(await page.evaluate(() => document.querySelector('#button-5').textContent)).toBe('clicked');
-      await page.click('#button-80');
-      expect(await page.evaluate(() => document.querySelector('#button-80').textContent)).toBe('clicked');
-    });
-    it('should double click the button', async({page, server}) => {
-      await page.goto(server.PREFIX + '/input/button.html');
-      await page.evaluate(() => {
-        window.double = false;
-        const button = document.querySelector('button');
-        button.addEventListener('dblclick', event => {
-          window.double = true;
-        });
-      });
-      const button = await page.$('button');
-      await button.click({ clickCount: 2 });
-      expect(await page.evaluate('double')).toBe(true);
-      expect(await page.evaluate('result')).toBe('Clicked');
-    });
-    it('should click a partially obscured button', async({page, server}) => {
-      await page.goto(server.PREFIX + '/input/button.html');
-      await page.evaluate(() => {
-        const button = document.querySelector('button');
-        button.textContent = 'Some really long text that will go offscreen';
-        button.style.position = 'absolute';
-        button.style.left = '368px';
-      });
-      await page.click('button');
-      expect(await page.evaluate(() => window.result)).toBe('Clicked');
-    });
-    it('should click a rotated button', async({page, server}) => {
-      await page.goto(server.PREFIX + '/input/rotatedButton.html');
-      await page.click('button');
-      expect(await page.evaluate(() => result)).toBe('Clicked');
-    });
     it('should select the text with mouse', async({page, server}) => {
       await page.goto(server.PREFIX + '/input/textarea.html');
       await page.focus('textarea');
@@ -330,16 +190,6 @@ module.exports.addTests = function({testRunner, expect}) {
       await page.mouse.up();
       expect(await page.evaluate(() => window.getSelection().toString())).toBe(text);
     });
-    it('should select the text by triple clicking', async({page, server}) => {
-      await page.goto(server.PREFIX + '/input/textarea.html');
-      await page.focus('textarea');
-      const text = 'This is the text that we are going to try to select. Let\'s see how it goes.';
-      await page.keyboard.type(text);
-      await page.click('textarea');
-      await page.click('textarea', {clickCount: 2});
-      await page.click('textarea', {clickCount: 3});
-      expect(await page.evaluate(() => window.getSelection().toString())).toBe(text);
-    });
     it('should trigger hover state', async({page, server}) => {
       await page.goto(server.PREFIX + '/input/scrollable.html');
       await page.hover('#button-6');
@@ -348,11 +198,6 @@ module.exports.addTests = function({testRunner, expect}) {
       expect(await page.evaluate(() => document.querySelector('button:hover').id)).toBe('button-2');
       await page.hover('#button-91');
       expect(await page.evaluate(() => document.querySelector('button:hover').id)).toBe('button-91');
-    });
-    it('should fire contextmenu event on right click', async({page, server}) => {
-      await page.goto(server.PREFIX + '/input/scrollable.html');
-      await page.click('#button-8', {button: 'right'});
-      expect(await page.evaluate(() => document.querySelector('#button-8').textContent)).toBe('context menu');
     });
     it('should set modifier keys on click', async({page, server}) => {
       await page.goto(server.PREFIX + '/input/scrollable.html');
@@ -389,12 +234,6 @@ module.exports.addTests = function({testRunner, expect}) {
       await page.keyboard.down('a');
       expect(await page.evaluate(() => window.lastEvent.repeat)).toBe(false);
     });
-    // @see https://github.com/GoogleChrome/puppeteer/issues/206
-    it('should click links which cause navigation', async({page, server}) => {
-      await page.setContent(`<a href="${server.EMPTY_PAGE}">empty.html</a>`);
-      // This await should not hang.
-      await page.click('a');
-    });
     it('should tween mouse movement', async({page, server}) => {
       await page.mouse.move(100, 100);
       await page.evaluate(() => {
@@ -422,25 +261,6 @@ module.exports.addTests = function({testRunner, expect}) {
       const button = await page.$('button');
       await button.tap();
       expect(await page.evaluate(() => getResult())).toEqual(['Touchstart: 0', 'Touchend: 0']);
-    });
-    it('should click the button inside an iframe', async({page, server}) => {
-      await page.goto(server.EMPTY_PAGE);
-      await page.setContent('<div style="width:100px;height:100px">spacer</div>');
-      await utils.attachFrame(page, 'button-test', server.PREFIX + '/input/button.html');
-      const frame = page.frames()[1];
-      const button = await frame.$('button');
-      await button.click();
-      expect(await frame.evaluate(() => window.result)).toBe('Clicked');
-    });
-    it('should click the button with deviceScaleFactor set', async({page, server}) => {
-      await page.setViewport({width: 400, height: 400, deviceScaleFactor: 5});
-      expect(await page.evaluate(() => window.devicePixelRatio)).toBe(5);
-      await page.setContent('<div style="width:100px;height:100px">spacer</div>');
-      await utils.attachFrame(page, 'button-test', server.PREFIX + '/input/button.html');
-      const frame = page.frames()[1];
-      const button = await frame.$('button');
-      await button.click();
-      expect(await frame.evaluate(() => window.result)).toBe('Clicked');
     });
     it('should type all kinds of characters', async({page, server}) => {
       await page.goto(server.PREFIX + '/input/textarea.html');
